requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
fpdf==1.7.2
scikit-learn==1.3.0
plotly==5.15.0
kagglehub==0.2.0 